Atual

+++ src/main/java/br/edu/avaliacao/models/AtribuicaoProfessor.java
@@ -1,12 +1,24 @@
 package br.edu.avaliacao.models;
 
 public class AtribuicaoProfessor {
-    private int id;
-    private int idTurma;
-    private int idProfessor;
+    private long id;
+    private long idTurma;
+    private long idProfessor;
+
     public AtribuicaoProfessor() {}
-    public AtribuicaoProfessor(int id, int idTurma, int idProfessor) { this.id = id; this.idTurma = idTurma; this.idProfessor = idProfessor; }
-    public int getId() { return id; } public void setId(int id) { this.id = id; }
-    public int getIdTurma() { return idTurma; } public void setIdTurma(int idTurma) { this.idTurma = idTurma; }
-    public int getIdProfessor() { return idProfessor; } public void setIdProfessor(int idProfessor) { this.idProfessor = idProfessor; }
+
+    public AtribuicaoProfessor(long id, long idTurma, long idProfessor) {
+        this.id = id;
+        this.idTurma = idTurma;
+        this.idProfessor = idProfessor;
+    }
+
+    public long getId() { return id; }
+    public void setId(long id) { this.id = id; }
+
+    public long getIdTurma() { return idTurma; }
+    public void setIdTurma(long idTurma) { this.idTurma = idTurma; }
+
+    public long getIdProfessor() { return idProfessor; }
+    public void setIdProfessor(long idProfessor) { this.idProfessor = idProfessor; }
 }
