+++ src/main/java/br/edu/avaliacao/models/Curso.java
@@ -1,10 +1,19 @@
 package br.edu.avaliacao.models;
 
 public class Curso {
-    private int id;
+    private long id;
     private String nome;
+
     public Curso() {}
-    public Curso(int id, String nome) { this.id = id; this.nome = nome; }
-    public int getId() { return id; } public void setId(int id) { this.id = id; }
-    public String getNome() { return nome; } public void setNome(String nome) { this.nome = nome; }
+
+    public Curso(long id, String nome) {
+        this.id = id;
+        this.nome = nome;
+    }
+
+    public long getId() { return id; }
+    public void setId(long id) { this.id = id; }
+
+    public String getNome() { return nome; }
+    public void setNome(String nome) { this.nome = nome; }
 }
