+++ src/main/java/br/edu/avaliacao/models/Resposta.java
@@ -1,16 +1,34 @@
 package br.edu.avaliacao.models;
 
 public class Resposta {
-    private int id;
-    private int idSubmissao;
-    private int idQuestao;
-    private Integer idOpcao;
+    private long id;
+    private long idSubmissao;
+    private long idQuestao;
+    private Long idOpcao;
     private String texto;
+
     public Resposta() {}
-    public Resposta(int id, int idSubmissao, int idQuestao, Integer idOpcao, String texto) { this.id = id; this.idSubmissao = idSubmissao; this.idQuestao = idQuestao; this.idOpcao = idOpcao; this.texto = texto; }
-    public int getId() { return id; } public void setId(int id) { this.id = id; }
-    public int getIdSubmissao() { return idSubmissao; } public void setIdSubmissao(int idSubmissao) { this.idSubmissao = idSubmissao; }
-    public int getIdQuestao() { return idQuestao; } public void setIdQuestao(int idQuestao) { this.idQuestao = idQuestao; }
-    public Integer getIdOpcao() { return idOpcao; } public void setIdOpcao(Integer idOpcao) { this.idOpcao = idOpcao; }
-    public String getTexto() { return texto; } public void setTexto(String texto) { this.texto = texto; }
+
+    public Resposta(long id, long idSubmissao, long idQuestao, Long idOpcao, String texto) {
+        this.id = id;
+        this.idSubmissao = idSubmissao;
+        this.idQuestao = idQuestao;
+        this.idOpcao = idOpcao;
+        this.texto = texto;
+    }
+
+    public long getId() { return id; }
+    public void setId(long id) { this.id = id; }
+
+    public long getIdSubmissao() { return idSubmissao; }
+    public void setIdSubmissao(long idSubmissao) { this.idSubmissao = idSubmissao; }
+
+    public long getIdQuestao() { return idQuestao; }
+    public void setIdQuestao(long idQuestao) { this.idQuestao = idQuestao; }
+
+    public Long getIdOpcao() { return idOpcao; }
+    public void setIdOpcao(Long idOpcao) { this.idOpcao = idOpcao; }
+
+    public String getTexto() { return texto; }
+    public void setTexto(String texto) { this.texto = texto; }
 }
